dribbleがstmありで動くように修正

+++ ros2_ws/src/robot_controller/dribble_controller/include/dribble_controller/dribble_controller.hpp
@@ -6,7 +6,7 @@
 
 #include "rclcpp/rclcpp.hpp"
 #include "std_msgs/msg/bool.hpp"
-#include "std_msgs/msg/float32.hpp"
+#include "std_msgs/msg/int16.hpp"
 
 class DribbleController : public rclcpp::Node
 {
@@ -21,14 +21,14 @@ private:
 
   bool is_configuration_valid_{true};
   bool dribble_enabled_{false};
-  double on_rpm_{600.0};
+  int on_rpm_{600};
   int command_period_ms_{10};
   int qos_depth_{1};
   std::string dribble_enabled_topic_;
   std::string dribble_rpm_topic_;
 
   rclcpp::Subscription<std_msgs::msg::Bool>::SharedPtr dribble_enabled_sub_;
-  rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr rpm_pub_;
+  rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr rpm_pub_;
   rclcpp::TimerBase::SharedPtr timer_;
 };
